Limpiado mierda y empezado disparo del enemigo hacia el jugador

+++ ProyectoII/Enemigo.h
@@ -23,7 +23,7 @@ private:
 	int vida;
 	const float tiempoBala = 1000;
 	Uint32 ultimaBala;
-	int angulo;
+	float angulo;
 
 };
